[refact] nand buffer -> command converter, command optimizer
