+++ SSD_Americano/SSD_Americano.vcxproj.filters
@@ -84,6 +84,18 @@
     <ClCompile Include="FlushCmd.cpp">
       <Filter>소스 파일\Command</Filter>
     </ClCompile>
+    <ClCompile Include="NandDriver.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
+    <ClCompile Include="CommandOptimizer.cpp">
+      <Filter>소스 파일\Nand</Filter>
+    </ClCompile>
+    <ClCompile Include="CommandTypes.h">
+      <Filter>헤더 파일\Nand</Filter>
+    </ClCompile>
+    <ClCompile Include="CommandConverter.cpp">
+      <Filter>소스 파일\Nand</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="EraseCmd.h">
@@ -109,6 +121,15 @@
     </ClInclude>
     <ClInclude Include="FlushCmd.h">
       <Filter>헤더 파일\Command</Filter>
+    </ClInclude>
+    <ClInclude Include="NandDriver.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="CommandOptimizer.h">
+      <Filter>헤더 파일\Nand</Filter>
+    </ClInclude>
+    <ClInclude Include="CommandConverter.h">
+      <Filter>헤더 파일\Nand</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
